Enable the tray icon
Enable the tray class of electron, by right clicking the icon, a menu with 2 options (show zettlr and quit) shall display. But the click event has not yet bind to these options.
However, the icon shall display before the main window opened. The icon would display after by the end of the initialize process of the application.